+++ pyobjc-framework-CoreAudio/PyObjCTest/test_audiodriverplugin.py
@@ -1,0 +1,10 @@
+from PyObjCTools.TestSupport import *
+
+import CoreAudio
+
+class TestAudioDriverPlugIn (TestCase):
+    def testFunctions(self):
+        self.fail("Add tests")
+
+if __name__ == "__main__":
+    main()